fix types

+++ src/sendall.c
@@ -1,7 +1,7 @@
 #include "../simple_socket.h"
 
 
-bool sendall(int32_t sock, void *buffer, uint32_t len,
+bool sendall(int32_t sock, void *buffer, int32_t len,
                 struct sockaddr *addr, socklen_t addr_len)
 {
     /* calls send() until the amount of bytes specified by
@@ -9,7 +9,7 @@ bool sendall(int32_t sock, void *buffer, uint32_t len,
 
     uint8_t *buffer_ptr = (uint8_t *)buffer;
     int16_t sent;
-    uint64_t total_sent = 0;
+    int32_t total_sent = 0;
 
     while (total_sent < len) {
         if ((sent = sendto(sock, buffer_ptr, len, 0,
